third_party/WebKit/LayoutTests/inspector/sources/debugger/async-operation-breakpoints.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/debugger-test.js"></script>
<script>

var dummy = function FAIL_should_not_pause_here() { return 0; };

function testFunction()
{
    debugger;

    setTimeout(dummy, 0);
    setTimeout(dummy, 30);
    (function setAsyncBreakpointForMe() { setTimeout(PASS_should_pause_here, 20); })();
    setTimeout(dummy, 0);
    Promise.resolve(42)
        .then(dummy);

    debugger; // Stop the debugger to receive pending AsyncOperation events.
}

function PASS_should_pause_here()
{
    return 0;
}

function test()
{
    var maxAsyncCallStackDepth = 4;
    InspectorTest.startDebuggerTest(step1, true);

    function step1()
    {
        InspectorTest.runTestFunctionAndWaitUntilPaused(step2);
    }

    function step2()
    {
        InspectorTest.debuggerModel.addEventListener(WebInspector.DebuggerModel.Events.AsyncOperationStarted, onAsyncOperationStarted);
        InspectorTest.DebuggerAgent.setAsyncCallStackDepth(maxAsyncCallStackDepth, step3);
    }

    function step3()
    {
        InspectorTest.waitUntilPausedAndPerformSteppingActions(["Resume", "Resume"], step4);
    }

    function step4()
    {
        InspectorTest.waitUntilPaused(didPause);
    }

    function didPause(callFrames, reason, breakpointIds, asyncStackTrace)
    {
        InspectorTest.captureStackTrace(callFrames);
        InspectorTest.addResult("Pause reason: " + reason);
        InspectorTest.completeDebuggerTest();
    }

    function onAsyncOperationStarted(event)
    {
        var operation = event.data.operation;
        var callFrame = operation.stackTrace && operation.stackTrace[0];
        var functionName = callFrame && callFrame.functionName;
        if (functionName === "setAsyncBreakpointForMe")
            InspectorTest.DebuggerAgent.setAsyncOperationBreakpoint(operation.id);
    }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests AsyncOperation breakpoints.
</p>
</body>
</html>
